platform/backend-ai/pyproject.toml
[tool.poetry]
name = "insty-backend-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.12"
pydantic = "^2.11.3"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.40"
sentry-sdk = "^2.26.1"
boto3 = "^1.38.18"
python-multipart = "^0.0.20"
langdetect = "^1.0.9"
psycopg = "^3.2.9"
openai = "^1.82.0"
moviepy = "^2.2.1"
pydub = "^0.25.1"
psutil = "^7.0.0"
jinja2 = "^3.1.6"
tiktoken = "^0.9.0"

# langchain-pinecone 라이브러리
pinecone-client = "^3.1.0"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
filetype = "^1.2.0"
celery = {extras = ["redis"], version = "^5.5.3"}
redis = "^4.6.0"
ffmpeg-python = "^0.2.0"
packaging = ">=23.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
